Map `long` and `Long` arguments to `BIGINT`
The SQL type `INTEGER` often (always?) maps to integer types which are only 32 bits (4 bytes) long while `BIGINT` maps to integer types which are 64 bits (8 bytes) long.

Considering this, mapping `long` and `Long` to `BIGINT` instead of `INTEGER` seems to make more sense.

FWIW, this is motivated by a bug we discovered in CockroachDB when sending `NULL` values which were sent as `NULL::INT4` by Jdbi 3.20.1 instead of `NULL::INT8` for a column declared as `INT8`.
https://github.com/cockroachdb/cockroach/issues/67605

+++ core/src/main/java/org/jdbi/v3/core/argument/PrimitivesArgumentFactory.java
@@ -28,7 +28,7 @@ class PrimitivesArgumentFactory extends DelegatingArgumentFactory {
         register(char.class, Types.CHAR, new ToStringBinder<>(PreparedStatement::setString));
         register(short.class, Types.SMALLINT, PreparedStatement::setShort);
         register(int.class, Types.INTEGER, PreparedStatement::setInt);
-        register(long.class, Types.INTEGER, PreparedStatement::setLong);
+        register(long.class, Types.BIGINT, PreparedStatement::setLong);
         register(float.class, Types.FLOAT, PreparedStatement::setFloat);
         register(double.class, Types.DOUBLE, PreparedStatement::setDouble);
     }
